+++ DataGenerator/Program.cs
@@ -401,7 +401,7 @@ namespace DataGenerator
                         {
                             //SkillInfo.AddUnreleasedMoveData();
                             //SkillInfo.AddMoveData();
-                            //SkillInfo.AddMoveData(887, 889);
+                            //SkillInfo.AddMoveData(845, 830);
                             //SkillInfo.AddMoveDataToAnims(120, 153);
                         }