+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 0.8.5 Changes #
 
 * Added Annihilape, Farigiraf, Samurott-H
-* Added Hyper Drill, Twin Beam, Rage Fist, Psyshield Bash, Stone Axe
+* Added Hyper Drill, Twin Beam, Rage Fist, Psyshield Bash, Stone Axe, Ceaseless Edge
 * PMDO is now updated to netcore 8.0!